Bug fix for issue #15
The CSV file for the "Full" export uses quotes if a field contains the
delimiter character (";") and also quotes inside the content will be
handled properly.
The apache commons-csv is now used to create the CSV instead of the
manual creation.

Refactored code, removed old unused code and updated deprecated methods
for the CSVFormat.